Remove build-time generated certificate
In its postinst script, the ssl-cert package generates a snakeoil
self-signed certificate that is made available so applications can work
with example configuration files out-of-the-box. In Docker images, this
implies that the key becomes public for all the users with access to the
image and that they all get the same key.

This patch removes the certificate and re-generates it in the entrypoint
of the image.

Signed-off-by: Athos Ribeiro <athos.ribeiro@canonical.com>